Move funções utilitárias para app de controle de tráfego
